This should get the client working to show the games and the ability to add games. Lots to clean up here.
