+++ client/src/App.tsx
@@ -1,26 +1,14 @@
 import { Box, List, MantineProvider } from "@mantine/core";
-import useSWR from "swr";
 import './App.css';
 import CreateGame from "./components/CreateGame";
 import '@mantine/core/styles.css';
-
-export interface Game {
-  id: number;
-  sport: string;
-  opponent: string;
-  location: string;
-}
-
-export const ENDPOINT = 'http://localhost:8080/api'
-
-const fetcher = (url: string) => fetch(`${ENDPOINT}/${url}`).then((r) => r.json())
+import { useGames } from "./services/queries";
 
 function App() {
-  const {data, mutate} = useSWR<Game[]>('games', fetcher)
+  const {data, mutate} = useGames();
 
-  console.log(data)
+  console.log(data);
 
-  //TODO: Iterate over the List and show the games
   return (
     <MantineProvider>
       <Box>
