+++ client/src/types/game.ts
@@ -1,0 +1,7 @@
+export interface Game {
+    id: number;
+    sport: string;
+    opponent: string;
+    location: string;
+  }
+
